Rules now work
